requirements.txt
google-api-python-client==2.142.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
langchain==0.2.12
langgraph==0.2.16
httpx==0.27.0
python-dotenv==1.0.1
pydantic==2.8.2
python-dateutil==2.9.0.post0
rich==13.7.1
